Update - intento agregar Selenium, Inicio examen 2009

+++ Unidad4POO/src/Banco/CuentaCorriente.java
@@ -64,7 +64,7 @@ public class CuentaCorriente {
 	}
 
 	//getters y setters
-	public float getSaldo() {
+	public double getSaldo() {
 		return saldo;
 	}
